+++ sdk/hdinsight/azure-mgmt-hdinsightcontainers/generated_samples/get_operations.py
@@ -35,6 +35,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2023-11-01-preview/examples/GetOperations.json
+# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2024-05-01-preview/examples/GetOperations.json
 if __name__ == "__main__":
     main()