+++ sdk/hdinsight/azure-mgmt-hdinsightcontainers/generated_samples/patch_cluster.py
@@ -110,6 +110,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2023-11-01-preview/examples/PatchCluster.json
+# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2024-05-01-preview/examples/PatchCluster.json
 if __name__ == "__main__":
     main()