+++ sdk/hdinsight/azure-mgmt-hdinsightcontainers/generated_samples/patch_ranger_cluster.py
@@ -64,6 +64,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2023-11-01-preview/examples/PatchRangerCluster.json
+# x-ms-original-file: specification/hdinsight/resource-manager/Microsoft.HDInsight/HDInsightOnAks/preview/2024-05-01-preview/examples/PatchRangerCluster.json
 if __name__ == "__main__":
     main()